15 物理引擎与物理组件 (#19)
* feat(PhysicsComponent): Add 管理 GameObject 物理属性的组件

* feat(PhysicsEngine): Add 数据成员和相关 getters & setters

* feat(PhysicsEngine): Add 注册、注销组件的方法

* feat(PhysicsComponent): Complete init() 和 clean() 的逻辑

* feat(PhysicsEngine): Add 更新所有注册的物理组件的核心循环方法

* fix(PhysicsEngine): Repair 二进制“*”:“glm::vec<2,float,glm::packed_highp>”不定义该运算符或到预定义运算符可接收的类型的转换

速度、力和质量都是 float 类型，但是更新方法中的时间却是 double 类型

* refactor: Modify all update()

传入的参数 deltaTime 类型从 double 变为 float

* refactor(FrameTimeController): Modify 所有与时间相关的数据成员类型从 float 到 double

* fix(GameApp): Repair 在函数调用中参数类型的 downcast

* chore(Renderer): Improve 默认参数的值类型为 double 而不是 float

* feat(GameApp): Add 管理物理引擎的数据成员和相应的初始化方法

* feat(Context): Add 物理引擎模块到 Context 类中并更新构造函数和相关 getter

* refactor(GameApp): Update 创建 Context 智能指针时的函数参数

* feat(SceneBase): Add 更新物理引擎逻辑到 update()

* test(GameScene): Check 物理组件是否工作正常

* refactor(GameApp): Move SceneManger 的清理逻辑

确保场景管理器在上下文销毁之前清理

+++ src/engine/component/parallax_component.h
@@ -49,7 +49,7 @@ public:
 protected:
     // 核心循环函数覆盖
     void init() override;
-    void update(double deltaTime, engine::core::Context&) override {} // 必须实现纯虚函数，留空
+    void update(float deltaTime, engine::core::Context&) override {} // 必须实现纯虚函数，留空
     void render(engine::core::Context& context) override;
 
 private:
