+++ src/engine/component/tilelayer_component.h
@@ -92,7 +92,7 @@ public:
 protected:
     // 核心循环方法
     void init() override;
-    void update(double deltaTime, engine::core::Context&) override {}
+    void update(float deltaTime, engine::core::Context&) override {}
     void render(engine::core::Context& context) override;
 
 private:
